change price data type db

+++ Models/Service.cs
@@ -5,7 +5,7 @@
         public string ServiceName { get; set; }
         public string Description { get; set; }
         public int StudioID {  get; set; }
-        public decimal Price { get; set; }
+        public float Price { get; set; }
         public string Image {  get; set; }
     }
 }
